Stop generating autograd functions for derivatives.yaml entries that only specify output differentiability.
the codegen for derivatives puts information about what is differentiable (inputs and outputs) together with the autograd function.
This is not a good idea, because it ties together the concept of "having an autograd function" with "specifying differentiability".

This results in a number of hacks.  One we fix here is wrt 'output_differentiability'.
Previously, we would generate autograd functions for entires that only specify output differentiability, e.g. values()/_values(),
and then warn that the actual autograd function was ignored.

Now, we avoid generating the autograd function and error out if we ever ignore an actual derivative specification.